Define helper for appending query params
In the case of "secret" Soundcloud tracks, the stream urls are given with a
`secret_token` query param, e.g.:

https://api.soundcloud.com/tracks/xxxxx/stream?secret_token=xxxx

The `play` function previously just appended `?client_id=...' to the end of the
stream URL for the track, so the resulting URL was incorrect:

https://api.soundcloud.com/tracks/xxxxx/stream?secret_token=xxxx?client_id=...

I defined a simple helper function for appending query params to a url that
checks to see if there is already any params defined.  This is currenly only in
use on the `play` function because that was my specific use case, but there
look to be some other parts of the code that could use it.